Feature/Account-Settings (#115)
* Basic Account functionality update

Adds a Member class object representing the user's Petlibro account, and makes modifications to accommodate it.
Allows the user to update some of their account settings, as well as show or use those account settings on the front-end with a sensor entity (lovelace, automations, scripts etc)
Can be easily built upon to include more settings and ways to manage things on your account in the future.

* Account icon for member entity

* test

* test

* Changed translation JSON keys to lowercase

* Account icon for member entity

* Changed symbols to unicode strings

---------

Co-authored-by: Jamie Jones <29973406+jjjonesjr33@users.noreply.github.com>